Add timed variant for each native function.
Swap the return order, so debug values are always the last element.

For now, the native functions will still calculate the time diff,
but will only return it if an extra (unused) argument exists.

In the future, this extra argument should be a proplist specifying
options, and the returned value could be a map or other type that
allows pattern matching over the debugged values.

A future optimization may be to only calculate the time difference
if the option argument is provided.