Combine `make` and `read` funcs. Add opt proplist.
This commit overhauls the BeamQN API. Instead of separate `make`
and `read` functions for every type, put everything in the same
NIF function and dynamically determine the encoding.

Advantages:
    - Significantly simpler API.
Disadvantages:
    - More type checking encoding and decoding.
    - Short running encoding or decoding has to be ran on the dirty
        scheduler, even if we don't expect it to be necessary.

In addition, this introduces an option proplist that will build an
option struct to determine which values should be returned.
In the future, it could be used to provide type suggestions.

Simplify test function names.